+++ mintapi/filters.py
@@ -4,7 +4,8 @@ General Filters helper classes
 
 from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass, field
-from typing import List
+from enum import Enum
+from typing import List, Optional, Union
 
 
 @dataclass
@@ -12,6 +13,14 @@ class MatchFilter(metaclass=ABCMeta):
     @abstractmethod
     def to_dict(self):
         pass
+
+
+@dataclass
+class AccountIdFilter(MatchFilter):
+    value: str
+
+    def to_dict(self):
+        return {"type": "AccountIdFilter", "accountId": self.value}
 
 
 @dataclass
@@ -63,6 +72,51 @@ class TagNameFilter(MatchFilter):
 
     def to_dict(self):
         return {"type": "TagNameFilter", "tagName": self.value, "exclude": True}
+
+
+@dataclass
+class DateFilter:
+    class Options(Enum):
+        """
+        Date options were inspected from the UI by clicking through all the available views
+        """
+
+        LAST_7_DAYS = 1
+        LAST_14_DAYS = 2
+        THIS_MONTH = 3
+        LAST_MONTH = 4
+        LAST_3_MONTHS = 5
+        LAST_6_MONTHS = 6
+        LAST_12_MONTHS = 7
+        THIS_YEAR = 8
+        LAST_YEAR = 9
+        ALL_TIME = 10
+        CUSTOM = 11
+
+    date_filter: Union[Options, str]
+    start_date: Optional[str] = None
+    end_date: Optional[str] = None
+
+    def __post_init__(self):
+        if isinstance(self.date_filter, str):
+            assert hasattr(self.Options, self.date_filter)
+        elif isinstance(self.date_filter, self.Options):
+            self.date_filter = self.date_filter.name
+        else:
+            raise ValueError(
+                "Date Filter must be one of allowed values in enum `DateFilter.Options"
+            )
+
+        if self.date_filter == self.Options.CUSTOM.name:
+            # validate required start and end dates
+            assert self.start_date is not None or self.end_date is not None
+
+    def to_dict(self):
+        filter_clause = {"dateFilter": {"type": self.date_filter}}
+        if self.date_filter == self.Options.CUSTOM.name:
+            filter_clause["dateFilter"]["startDate"] = self.start_date
+            filter_clause["dateFilter"]["endDate"] = self.end_date
+        return filter_clause
 
 
 @dataclass