+++ mintapi/signIn.py
@@ -59,7 +59,7 @@ MFA_METHODS = [
     },
     {
         constants.MFA_METHOD_LABEL: constants.MFA_VIA_SMS,
-        SELECT_CSS_SELECTORS_LABEL: "#ius-mfa-sms-otp-card-challenge",
+        SELECT_CSS_SELECTORS_LABEL: '#ius-mfa-sms-otp-card-challenge, [data-testid="challengePickerOption_SMS_OTP"]',
         INPUT_CSS_SELECTORS_LABEL: "#ius-mfa-confirm-code",
         SPAN_CSS_SELECTORS_LABEL: '[data-testid="VerifyOtpHeaderText"], #VerifyOtpHeader',
         BUTTON_CSS_SELECTORS_LABEL: '#ius-mfa-otp-submit-btn, [data-testid="VerifyOtpSubmitButton"]',
@@ -120,8 +120,13 @@ def get_email_code(imap_account, imap_password, imap_server, imap_folder, delete
             if not re.search("do_not_reply@intuit.com", frm, re.IGNORECASE):
                 continue
 
-            if not re.search("Your Mint Account", subject, re.IGNORECASE):
+            p = re.search(r"(\d\d\d\d\d\d) Mint code", subject)
+            if p:
+                code = p.group(1)
+            elif not re.search("Your Mint Account", subject, re.IGNORECASE):
                 continue
+            else:
+                code = ""
 
             date_tuple = email.utils.parsedate_tz(msg["Date"])
             if date_tuple:
@@ -138,13 +143,15 @@ def get_email_code(imap_account, imap_password, imap_server, imap_folder, delete
 
             logger.debug("DEBUG: EMAIL HEADER OK")
 
-            body = next(msg.walk()).get_payload(None, True).decode()
-
-            p = re.search(r"Verification code:<.*?(\d\d\d\d\d\d)\b", body, re.S | re.M)
-            if p:
-                code = p.group(1)
-            else:
-                logger.error("FAIL1")
+            if code == "":
+                body = next(msg.walk()).get_payload(None, True).decode()
+                p = re.search(
+                    r"Verification code:<.*?(\d\d\d\d\d\d)\b", body, re.S | re.M
+                )
+                if p:
+                    code = p.group(1)
+                else:
+                    logger.error("FAIL1")
 
             logger.debug("DEBUG: CODE FROM EMAIL:", code)
 
@@ -371,6 +378,7 @@ def sign_in(
 
         # Wait until logged in, just in case we need to deal with MFA.
 
+        handle_login_failures(driver)
         if not bypass_verified_user_page(driver):
             # if bypass_verified_user_page was present, then MFA already done
             bypass_passwordless_login_page(driver)
@@ -466,6 +474,49 @@ def handle_different_page_username_password(driver, email):
             if username_element.text == email:
                 username_element.click()
                 break
+
+
+def handle_login_failures(driver):
+    try:
+        WebDriverWait(driver, 0).until(
+            expected_conditions.presence_of_element_located(
+                (
+                    By.XPATH,
+                    '//div[contains(text(), "We can\'t find anyone with ")][@id="ius-identifier-first-error"]',
+                )
+            )
+        )
+        raise RuntimeError(
+            "Login to Mint failed: Mint does not recognize your login email"
+        )
+    except TimeoutException:
+        pass
+
+    try:
+        WebDriverWait(driver, 0).until(
+            expected_conditions.presence_of_element_located(
+                (
+                    By.XPATH,
+                    "//div[contains(text(), 'The password you entered is incorrect.')]",
+                )
+            )
+        )
+        raise RuntimeError("Login to Mint failed: incorrect password")
+    except TimeoutException:
+        pass
+
+    try:
+        WebDriverWait(driver, 0).until(
+            expected_conditions.presence_of_element_located(
+                (
+                    By.ID,
+                    "RecaptchaHeader",
+                )
+            )
+        )
+        raise RuntimeError("Login to Mint failed: Captcha presented")
+    except TimeoutException:
+        pass
 
 
 def bypass_verified_user_page(driver):
@@ -657,29 +708,41 @@ def submit_mfa_code(mfa_token_input, mfa_token_button, mfa_code):
 def account_selection_page(driver, intuit_account):
     # account selection screen -- if there are multiple accounts, select one
     try:
-        select_account = driver.find_element(By.ID, "ius-mfa-select-account-section")
-        if intuit_account is not None:
-            account_input = select_account.find_element(
-                By.XPATH,
-                "//label/span[text()='{}']/../preceding-sibling::input".format(
-                    intuit_account
-                ),
-            )
-
-            account_input.click()
         WebDriverWait(driver, 20).until(
             expected_conditions.presence_of_element_located(
                 (
                     By.CSS_SELECTOR,
-                    "[data-testid='SelectAccountContinueButton']",
+                    '[data-testid="SelectAccountForm"]',
                 )
             )
         )
-        mfa_code_submit = driver.find_element(
+        select_account = driver.find_element(
+            By.CSS_SELECTOR, '[data-testid="SelectAccountForm"]'
+        )
+        if intuit_account is not None:
+            account_input = select_account.find_element(
+                By.XPATH,
+                "//*/span[text()='{}']/../../../preceding-sibling::input".format(
+                    intuit_account
+                ),
+            )
+            # NOTE: We need to execute a script because simply using account_input.click()
+            #       results in ElementClickInterceptedException.
+            driver.execute_script("arguments[0].click()", account_input)
+
+        WebDriverWait(driver, 20).until(
+            expected_conditions.presence_of_element_located(
+                (
+                    By.CSS_SELECTOR,
+                    "#ius-sign-in-mfa-select-account-continue-btn, [data-testid='SelectAccountContinueButton']",
+                )
+            )
+        )
+        driver.find_element(
             By.CSS_SELECTOR,
             '#ius-sign-in-mfa-select-account-continue-btn, [data-testid="SelectAccountContinueButton"]',
         ).click()
-    except NoSuchElementException:
+    except (TimeoutException, NoSuchElementException):
         logger.info("Not on Account Selection Screen")