+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 --> NOTE! The 1.x release series only supports and contain fixes for the previous Mint UI.
     New feature work occurs in the 2.x release series and only supports the new Mint UI / endpoints.
     Please see the README for more information.
+    
+2.8
+---
+- handle new MFA email format w/ code in subject (#538)
+- update for new SMS MFA UI (#536)
+- improve README for creating and passing in your own driver (#542)
+
+2.7
+---
+- fix for Intuit account selection login bug (#530)
+- Add AccountIdFilter to Endpoint Filters (#526)
+
+2.6
+---
+- add support for accessing the Trend API endpoint (#480)
+- bump selenium-requests dependency version to resolve issues (#521)
+- attempt to show the reason for a login failure (#523)
 
 2.5
 ---